+++ tests/unit-core/test-number-converter.c
@@ -13,8 +13,7 @@
  * limitations under the License.
  */
 
-#include "ecma-globals.h"
-#include "ecma-helpers.h"
+#include "jerryscript.h"
 
 #include "test-common.h"
